README: Why to prefer dev build, how to switch
It should be stated what build is used by default and provide a justification for it. A newcomer would be likely confused by why we use dev build in production, so I have copied the justification from #50 and referred to it regarding info about how to switch. I suppose this is going to evolve but we should document the current status.