FIX: Tests for asMutable fail on some platforms
Regarding #104

Test if type of value passed to asDeepMutable is object and only recur if it is. Otherwise return unchanged value. This shall prevent calling asMutable with deep option on *leaf* properties of immutable object.

Also the if statement is now split into multiple lines for readability - there are four alternative conditions there and it was looking really messy on a single line IMO.